+++ src/m_menu.c
@@ -3251,7 +3251,7 @@ bool M_Responder(event_t *ev)
         if (!messagetoprint)
         {
             // select previous menu item
-            if (ev->data1 > 0)
+            if (ev->data2 > 0)
             {
                 key = KEY_UPARROW;
                 mousewait = I_GetTime() + 3;
@@ -3259,7 +3259,7 @@ bool M_Responder(event_t *ev)
             }
 
             // select next menu item
-            else if (ev->data1 < 0)
+            else if (ev->data2 < 0)
             {
                 key = KEY_DOWNARROW;
                 mousewait = I_GetTime() + 3;